feat: update file structure
